requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
flask-restx==1.2.0
pymongo==4.3.3
python-dotenv==1.0.0
PyJWT==2.7.0
gunicorn==20.1.0
flask-pymongo==2.3.0
gunicorn>=20.1.0
flask-cors